requirements.txt
streamlit==1.35.0
pypdf==4.2.0
PyPDF2==3.0.1
langchain==0.2.6
openai==1.35.9
openpyxl==3.1.4
python-dotenv==1.0.1
streamlit==1.35.0
pandas==1.5.3
numpy==1.26.4